[BE][Fix] 기념일 카테고리 고유번호 타입 변경
Long -> Integer

+++ BE/src/main/java/com/d201/fundingift/funding/repository/AnniversaryCategoryRepository.java
@@ -3,5 +3,5 @@ package com.d201.fundingift.funding.repository;
 import com.d201.fundingift.funding.entity.AnniversaryCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnniversaryCategoryRepository extends JpaRepository<AnniversaryCategory, Long> {
+public interface AnniversaryCategoryRepository extends JpaRepository<AnniversaryCategory, Integer> {
 }
